changed the db-timestamp type to bigint, and trying to fill the database with data. But it is too fast right now and db can't keep up

+++ backend/create_database.sql
@@ -4,71 +4,71 @@ USE StromDaten;
 
 CREATE TABLE IF NOT EXISTS BraunkohleErzeugung (
 ID INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
-Timestamp_Unix INT,
+Timestamp_Unix BIGINT,
 Menge INT);
 
 CREATE TABLE IF NOT EXISTS KernEnergieErzeugung (
 ID INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
-Timestamp_Unix INT,
+Timestamp_Unix BIGINT,
 Datum_Zeit VARCHAR(100),
 Menge INT);
 
 CREATE TABLE IF NOT EXISTS OffshoreWindErzeugung(
 ID INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
-Timestamp_Unix INT,
+Timestamp_Unix BIGINT,
 Menge INT);
 
 CREATE TABLE IF NOT EXISTS WasserkraftErzeugung (
 ID INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
-Timestamp_Unix INT,
+Timestamp_Unix BIGINT,
 Menge INT);
 
 CREATE TABLE IF NOT EXISTS Erzeugung_Sonstige_Konventionell (
 ID INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
-Timestamp_Unix INT,
+Timestamp_Unix BIGINT,
 Menge INT);
 
 CREATE TABLE IF NOT EXISTS Erzeugung_Sonstige_Erneuerbar (
 ID INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
-Timestamp_Unix INT,
+Timestamp_Unix BIGINT,
 Menge INT);
 
 CREATE TABLE IF NOT EXISTS BiomasseErzeugung (
 ID INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
-Timestamp_Unix INT,
+Timestamp_Unix BIGINT,
 Menge INT);
 
 CREATE TABLE IF NOT EXISTS OnshoreWindErzeugung (
 ID INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
-Timestamp_Unix INT,
+Timestamp_Unix BIGINT,
 Menge INT);
 
 CREATE TABLE IF NOT EXISTS PhotovoltaikErzeugung (
 ID INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
-Timestamp_Unix INT,
+Timestamp_Unix BIGINT,
 Menge INT);
 
 CREATE TABLE IF NOT EXISTS SteinkohleErzeugung (
 ID INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
-Timestamp_Unix INT,
+Timestamp_Unix BIGINT,
 Menge INT);
 
 CREATE TABLE IF NOT EXISTS PumpspeicherErzeugung (
 ID INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
-Timestamp_Unix INT,
+Timestamp_Unix BIGINT,
 Menge INT);
 
 CREATE TABLE IF NOT EXISTS ErdgasErzeugung (
 ID INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
-Timestamp_Unix INT,
+Timestamp_Unix BIGINT,
 Menge INT);
 
 CREATE TABLE IF NOT EXISTS Verbrauch_Gesamt (
 ID INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
-Timestamp_Unix INT,
+Timestamp_Unix BIGINT,
 Menge INT);
 
 CREATE TABLE IF NOT EXISTS Verbrauch_Pumpspeicher (
 ID INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
-Timestamp_Unix INT,
+Timestamp_Unix BIGINT,
 Menge INT);
